Declare `hfs` as PUBLIC -fno-strict-aliasing.
hfs makes extensive use of pointer-cast-based type punning. This is undefined behavior due to the strict aliasing rule. We must shut off the strict aliasing rule for this codebase, shutting off potential compiler optimizations. This incurs a performance cost, but everything this library does is likely to be HDD I/O bound anyway from a performance standpoint, and the optimizations it shuts off are the exact ones that can go horribly wrong in the presence of illegal-cast-based type punning, so it's our best option.

HFSPlusCatalogRecord and related types caught my eye here, but I'm not confident there's not more.  HFSPlusAttrRecord, conversely, stands out as doing it correctly -- it's undefined behavior in C++, but union-based type punning is explicitly part of the C standard.