Wrap "check and cast" pattern for HFSPlusCatalogFolder.
If we use this literally everywhere, it might make it easier to convert to a union-based approach that does not violate the strict aliasing rule. That is probably unreasonably ambitious. However, I would rather have these helpers available instead of repeating the check everywhere, and am likely to rewrite the pattern where I see it, especially if I see repeated casts of the same (ambiguous) pointer, so here it is for my personal convenience.