Switch from strlcpy to strlen + memcpy.
strlcpy, while convenient, is a BSD API not available on the Debian image used for the CI workflow (although the Ubuntu image specified in the Dockerfile is perfectly fine with it). Since strncpy does not provide any way to distinguish between "okay, I copied your entire string" and "btw I couldn't find room for the NUL terminator, hope you didn't need that, lul", I'm using strlen+memcpy explicitly, which is at least more obvious about how it will not help you maintain C string validity in any way.

This implementation relies on the original string to have a \0 in the expected spot, but so does strlen (and all other str-based APIs) so it's not like this makes the situation worse.