Rename symlink policy "copy" to "clone_link".
When writing the tests, I kept getting myself confused because I kept expecting "copy" to make a copy of the target file, which is the exact opposite of what it does (it copies the link itself instead). If I get confused using it, and I'm the one who *wrote* it, it's almost certainly going to be confusing to everyone who tries to use it in the future, too. `clone_link` is much less ambiguous about what is happening.

I considered `link` but I worried it might make people think that it would create a link to the same destination, which is not quite what happens -- it creates a literally identical link, which means a relative link will be relative to its new location.  This is the best name I can think of for intuitively matching these semantics.

Separately, I am not thrilled with the flag name `--special_modes` but I keep not coming up with any better ideas.