Fix hfsplus.h indentation. Delete tryCatalogRecordAsThread family.
Nothing handles those cases anyway, the helper is just clutter. It can easily be added later if needed. (also the underlying file got removed in a catalog.c cleanup that was a paste from a version of the repo where I didn't add the extraneous helpers in the first place, oops)